SI-8323 Erase constant types for better double-def checks
Historically, Constant Types in method signatures have not been
erased. Rather, this was done in the backend.

This state of affairs leads to wrong decisions when the erasure
phase makes decisions about whether a pair of methods will clash
in bytecode.

This commit changes `ErasureMap` to take the underlying type of a
constant type.

This change obviates the fix for SI-4753, dcc07bd9f.

I've added test to show that constants are still inlined.

I believe this will answer the question in the backend about why
existential types make it that far; they come from the underlying
of a constant type representing a class literal. I'll leave it to
Lukas to verify that the code in question is now dead and to remove
it in his backend work.

I searched through the post-erasure phases to find things that
we should adapt.

   posterasure  -
      lazyvals  -
    lambdalift  -
  constructors  x elides fields of constant types
       flatten  -
         mixin  x disable checkinit for constant fields
                  derives setter/getter symbols of fields
                  elides setters for constant fields
                  uses a literal as the body of for getters for const
                  fields
       cleanup  - looks at constant trees for scala.Symbol interning,
                  but not at types
    delambdafy
     (backend)  x unwrap constant types in `toTypeKind`. This will
                  become dead code, and I'll remove in a subsequent
                  commit.

Adapting these spots is pretty straight forward:

  - We use something like `enteringErasure(sym.info.isConstantType)`
    rather than looking that
  - When deriving getter and setter symbols, we follow the same
    "clone before erasure" discipline as we do for other mixed-in
    members.