SI-8907 Avoid over elaborating symbol infos
In 0ccdb151f, `isModuleNotMethod` was modified to force the info
of the symbol if the current phase was post-refchecks before
reading flags.

This was done to let the refchecks info transformer run. That
transformer introduces the `lateMETHOD` flag on a symbol originating
from a nested object, which has a split personality after that phase
as the module *and* the module accessor method. Reading the flags
without forcing the info was prone to data races.

However, after this change, we might also run *later* info
transformers on a symbol. In the enclosed test case, or the reported
bug from which it is extracted, this led to us unnecessarily
elaborating the signature for a method that we don't refer to
in proram we are compiling. Under certain circumstances involving
dependent method types, this led to us tripping a
`MissingRequirementError` when we eventually asked for the `.info`
of a `StubSymbol` representing an absent class.

`StubSymbol`s exist to defer compilation failure when we
encounter a reference to a class that is absent from the
current compilation classpath. The principle is: if you don't
need a particular method from a class in a library your are
using, the compiler should not fail if some type referenced
in its signature is absent. This behaviour is modelled after
javac.

The enclosed test follows an established pattern for exercising
our stubs; compile, delete classes, then compile client code.