SI-8934 Fix whitebox extractor macros in the pres. compiler
The code that "aligns" patterns and extractors assumes that it can
look at the type of the unapply method to figure the arity of the
extractor.

However, the result type of a whitebox macro does not tell the
whole story, only after expanding an application of that macro do
we know the result type.

I'm not quite sure how this hasn't caused a problem in the batch
compiler, but in the presentation compiler we end up with `Error-`
and `NoType`-s floating around the code paths in `unapplyMethodTypes`
and `alignPatterns.apply`.

This commit sharpens the final result type of the macros method
type with the type of the expanded macro before performing the arity
analysis.

The enclosed test used to fail with:

    too many patterns for trait api: expected 1, found 2

But now passes.