SI-1415 Fix overriding w/ qualified private over and public under
Since 22a098bf7, qualified private methods, which are private within
a non-package, have been name mangled during the super accessors
transform.

However, this means that an override of such a member will not
be seen to match it unless it has the same access qualification.
In the enclosed test case, refchecks failed with:

    object creation impossible, since method baz in trait Bar of type => Unit is not defined

At that point, `findMember`, sees *both* `baz` and `Foo$$baz` as
members.

This commit uses the standard `makeNotPrivate` method instead of the
lower level.