SI-7243 Fix reification of TypeTreeWithDeferred...
The reifier has a default strategy of generating a call
`u.Foo(...)` when reifying a case class with product prefix
`Foo`. This works well for all of the AST nodes visible in
the public reflection API, but falls down for an extra one
added in the compiler layer of the cake,
`TypeTreeWithDeferredRefCheck`.

This commit handles that one specifically, by unwrapping the embedded
`TypeTree` and reifying that instead.