Extremely hacky workaround for sbt/sbt#1872
We hack the `update` task to think its called from a project with a
different (bogus) name than it really is. For example, when we call
update in `library` subproject (that has name `scala-library`) but update
task sees this call coming from an project named `scala-library-bogus`.

I hope we'll find some other fix for sbt/sbt#1872 in the future.