SI-7187 Avoid spurious warning for explicit eta-expansion
`typedEta` was offloading the work of creating a function
to `adapt`, which we have changed to warn when the expected
type is `Function0`. This led to a spurious warning for:

```
object Test {
  def testC00()(): Int = 1
  val testC00R = testC00 _
}
```

(This snippet is from run/byname.scala.)

This commit changes `typedEta` to diretly call the relevant logic
from within `adapt`. To facilitate this, I have extracted it to
`EtaExpansion`.