SI-6962 Fix crasher with erroneous code, defaults
When naming a `DefDef`, `setPrivateWithin` typechecks the qualifier
in the access modifier. If this does not designated an enclosing
class, an error is issued and `Infer.setError` is used to mark
the `DefDef` as erroneous.

This was done by a) setting the type to `ErrorType` and b) setting
the tree's `Symbol` to a new error symbol.

Later, the type completer for this tree sees only this error symbol,
rather than the actual symbol for the method. This thwarts synthesis
of default getters when walking up the context chain looking for the
scope that contains the method symbol, and an assertion is tripped.

The easy way out would by to disable all or part of the method
type completer when the `DefDef`'s has an error symbol.

But this seems unnecessarily ...