Avoid unwanted boostrap JARs in the subproject classpaths
I added build configuration to ask SBT to forcible override the Scala
version in the subproject to the declared bootstrap version. I did
this in response to a warning by SBT the declared `scalaVersion`
has been evicted.

However, we actually *want* the newer version. To get rid of the
warning, we need to exclude all `"org.scala-lang" % "*"` when
depending on Scala modules. This cleanly breaks the cycle.

Here's a diff of `*/dependencyClasspath` as a result of this patch:

    https://gist.github.com/retronym/217c76001b1b81798042