SI-9319 Avoid "filename too long" errors for anon functions
Under -Ydelambdafy:method, the name of the anonymous class is
derived from the names of the original enclosing class and
term of the function.

However no effort is made to conform to the `-Xmaxclassname` setting
when constructing the class name from the components.

Recently, in 029cce7, I changed uncurry to selectively fallback
to the old method of emitting lambdas when we detect that
`-Ydelambdafy:method` would subvert specialization.

As a result of this, we could now have a delambdafy translated
function enclosed in an anonymous class representing an old-style
function. If that function, in turn, is heavily nested, its long
name now contributes a long suffix to the delambdafy class's name.

This led to a "filename too long" error in `run/kmpSliceSearch.scala`.
I have extract the essence of the failure in `pos/t9319.scala`.

I then changed delambdafy to use the unexpanded names of the owners
rather than the full name. This makes things work as expected.

I haven't further protected against pathological cases, given
that in 2.12.x and indylambda, we won't ever emit these classes.

A test for the names of delambdafy functions was modified to account
for this change.