Improve performance of Constant stringification
Shapeless uses a few tricks to let constant types flow into type comparisons.
This exposed the poor performance of `Constant#toString`. `toString` on
types ought to be cheap, because under the current compiler architecture, we
pay a cost for error message creation during what might end up to be a
successful implicit search.

This commit avoids the expensive `stringValue flatMap f`, in favour
of explicit threading of a SringBuffer.

Thanks to Rory Graves for pinpointing the issue.