[presip] Allow a mix of inferred and provided type arguments.
The path of least resistance for experimentation was to use a
bacticked `_` as the wildcard specifier. This isn't a suggestion
for the final syntax. A naked `_` is one possibility, a magic
type alias like `Predef.?` is another.

`TypeApply(fun, targs)` which contains one or more `_`-s as
in `targs` is rewritten to `fun` with its `PolyType` constained
with the knowledge of the explicit type arguments: we create a
synthetic type parameter that is upper and lower bounded by
the provided type arg.

I've only catered for wildcards that are used directly as type
arguments, no support (yet) for `test[A, Set[<inferme>]]`. That
would demand rather a different approach from what I've done here.

See the TODOs in the code for some questions to answer and work to do.

Demo:

```
class Test {
  def main(args: Array[String]): Unit = {
    def test[A, B](a: A): B = ???

    test[String, Int]("")
    test[`_`   , Int]("")

    type \/[A, B] = Either[A, B]
    import collection.immutable.Set
    val nums: Set[Int] = null

    val a = nums.flatMap[String \/ Int, Set[String \/ Int]](a => Set(Right(a))) // Note required duplication of String \/ Int argument

    val b = nums.flatMap[String \/ Int, `_`               ](a => Set(Right(a))) // Note required duplication of String \/ Int argument

    def test2[A, B >: A] = ???
    test2[Int, Int]
    test2[Int, `_`]
  }
}

/**
  test[String, Int]("");
  test[String, Int]("");
  type \/[A, B] = Either[A,B];
  import scala.collection.immutable.Set;
  val nums: scala.collection.immutable.Set[Int] = null;
  val a: scala.collection.immutable.Set[\/[String,Int]] = nums.flatMap[\/[String,Int], scala.collection.immutable.Set[\/[String,Int]]](((a: Int) => scala.collection.immutable.Set.apply[scala.util.Right[Nothing,Int]](scala.`package`.Right.apply[Nothing, Int](a))))(immutable.this.Set.canBuildFrom[\/[String,Int]]);
  val b: scala.collection.immutable.Set[\/[String,Int]] = nums.flatMap[\/[String,Int], scala.collection.immutable.Set[\/[String,Int]]](((a: Int) => scala.collection.immutable.Set.apply[scala.util.Right[Nothing,Int]](scala.`package`.Right.apply[Nothing, Int](a))))(immutable.this.Set.canBuildFrom[\/[String,Int]]);
  def test2[A, B >: A]: Nothing = scala.this.Predef.???;
  test2[Int, Int];
  test2[Int, Int]
**/

```