Better support for classOf in presentation compiler
Currently, the tree `Predef.classOf[Foo]` or `classOf[Foo]` is
eliminated from the tree by compilation, leaving a literal in its
place.

This commit uses an existing configuration hook in the compiler
to disable this class constant folding under the presentation
compiler.

The treatment and encoding of `classOf` is a bit of a mess for
other areas, too:

   - https://issues.scala-lang.org/browse/SI-8384
   - comments in `TreeGen#mkClassOf`

Longer term, we might try to convert `classOf` to a macro that always
returns a tree like the one I'm using in this commit, updating the
compiler backend as needed to avoid actually calling `Predef.classOf`
at runtime.