Implement indy-structural without using dynalink
To start with, use a custom bootstrap method that instantiates
a class that (softlye) holds a method cache, and contains the
logic to perform the invocations.

The actual invocation is still done with java.lang.reflect.

I expect that this approach ought to be on par with the status quo
performance wise.