More fixes based on feedback by Lukas
Crucially, the fully-defined expected type must be checked for
conformance to the original expected type!!

The logic in adaptToSam that checks whether pt is fully defined
probably needs some more thought. See pos/t8310 for a good test case.
Argument type checking is a challenge, as we first check against a lenient pt
(this lenient expected type has wildcards, and thus is not fully defined,
 but we should still consider sam adaptation a success even if we end up with
 wildcards for some unknown type parameters, they should be determined later).