Test that SAM conversion happens after implicit view application
A function node is first type checked, and parameter types are inferred,
regardless of whether the expected function type is one of our
built-in FunctionN classes, or a user-defined Single Abstract Method type.

`typedFunction` always assigns a built-in `FunctionN` type to the tree, though.

Next, if the expected type is a (polymorphic) SAM type, this creates a tension
between the tree's type and the expect type. This gap is closed by the adapt method,
by applying one of the implicit conversion in the spec in order
(e.g., numeric widening, implicit view application, and now, also SAM conversion)

Thus, `adaptToSam` will assign the expected SAM type to the `Function` tree.
(This may require some type inference.) The back-end will emit the right
invokedynamic instruction that uses Java's LambdaMetaFactory to spin up a
class that implements the target method (whether it's defined in
FunctionN or some other Java functional interface).