SD-229 Make DelayedInit transform create/use param accessor fields
Rather than moving a reference to a constructor parameter into
`delayedEndpoint` and incurring a compiler crash.

The test in pending suggests to me that we might be better of fixing
this in a different way: we could make `LambdaLift` aware of
DelayedInit and have it create/use param accessor fields as proxies
for references in delayed init template bodies.