[perf] Simplify synthetic case class equals for top-level classes
We have been emmiting a type pattern in order to check not
only the runtime class, but also the prefix. This is semantically
equivalent to an `isInstanceOf` check for non-member case classes.

This commit special cases these (as they are common!) to just
emit `isInstanceOf` and avoid later translation steps in
the pattern matcher.