SI-10180 Scan generic signatures when populating InnerClass attribute
Generic signatures, like descriptors, include class names directly into the
grammar, rather than referring to class constants in the constant pool.

For the benefit of other compilers and tools consuming generic signatures,
we should record the metadata for inner classes that the reference.

This commit includers a parser for generic signatures that is designed for
performance with the simplifying assumption that it need only read
internal names. I started with the existing parser in ClassfileParser
that builds up symbols/types based on generic signatures. I don't see
an easy way to share more code between the two parsers without compromising
performance.

The new parser is used in the backend to trawl for referenced names in
the signatures of class definitions, field types, and method types.