Reduce overhead of lambda deserialization
For classes that host more than 251 lambdas, the previous commit
introduced a sequence of calls to LambdaDeserializer instances,
each handling a group of 251 target methods. Control flows is
handled with exceptions.

This commit supresses the exception for the IllegalArgumentException
that is thrown when a SerializedLambda is not a member of the current
group.