Separate shell from compiler, reduce mutation
The shell (ILoop) should only depend on the repl (IMain),
and not on the rest of the compiler. Working towards a
stable API that multiple frontends could use to provide
a shell on top of a repl-oriented compiler (in the same
vein as IDEs built on the presentation compiler).

Move config/settings out of package object, and into
ShellConfig / ReplReporterConfig.

IMain should not do any reporting to the console,
just send strings / trees to frontend, which will
then send it to output (colorizing where required).

TODO:
  - bring back color output by parsing trees received
    by shell from imain

Misc:
  - Rename unwrap to rootCause
  - Direct way to make welcome message reproducible for tests
  - paste prints a bit more prompt