Avoid disabling statistics
This commit avoids disabling statistics to avoid the JVM to unstabilize
all the time and to prevent misbehaviours in concurrent builds recycling
the same classloader.

The main problem here is that if one global has statistics enabled and
the second one doesn't, our logic of `enabled` was setting the
pseudo-static to false, preventing the correct recording of statistics
in the global that does want to record them. Now, the logic to disable
these pseudo-statics when `statistics = false` (which is the case almost
always), has been removed.

TL;DR: Once a global enables statistics, the rest has to pay the price
of a 2/3% performance hit. There is no going back.